refactor(ast): organize printer
